+++ r_main.h
@@ -30,10 +30,10 @@
 extern fixed_t		viewcos;
 extern fixed_t		viewsin;
 
-extern int32_t		viewwidth;
-extern int32_t		viewheight;
-extern int32_t		viewwindowx;
-extern int32_t		viewwindowy;
+extern int16_t		viewwidth;
+extern int16_t		viewheight;
+extern int16_t		viewwindowx;
+extern int16_t		viewwindowy;
